+++ src/components/project/task/milestone/MilestoneCardMenu.tsx
@@ -3,6 +3,8 @@ import React, {Fragment} from 'react';
 import {IoEllipsisVertical} from "@react-icons/all-files/io5/IoEllipsisVertical";
 import {Menu, Transition} from "@headlessui/react";
 import {classNames} from "@/utils/common";
+import {useRecoilValue} from "recoil";
+import {projectUserAuthStateStore} from "@/store/project/crews/ProjectUserAuthStateStore";
 
 interface MilestoneCardMenuProps {
     milestoneId: bigint;
@@ -11,6 +13,7 @@ interface MilestoneCardMenuProps {
 }
 
 function MilestoneCardMenu({milestoneId, onEditClickHandler, onDeleteClickHandler}: MilestoneCardMenuProps) {
+    const authInfo = useRecoilValue(projectUserAuthStateStore);
 
     const milestoneMenus = [
         {
@@ -26,55 +29,59 @@ function MilestoneCardMenu({milestoneId, onEditClickHandler, onDeleteClickHandle
     ]
 
     return (
-
-        // <div className="flex-shrink-0 pr-2 overflow-visible">
-        <Menu as="div" className="self-start flex-shrink-0 pr-2 text-center">
-            {/*<Menu as="div" className="relative inline-block text-center">*/}
-            <div>
-                <Menu.Button
-                    className="inline-flex h-8 w-8 items-center justify-center rounded-full bg-transparent bg-white text-gray-400 hover:text-gray-500 focus:outline-none">
-                    <span className="sr-only">마일스톤 메뉴</span>
-                    <IoEllipsisVertical className="h-5 w-5" aria-hidden="true" data-role='milestone-menu'/>
-                </Menu.Button>
-            </div>
-            <Transition
-                as={Fragment}
-                enter="transition ease-out duration-100"
-                enterFrom="transform opacity-0 scale-95"
-                enterTo="transform opacity-100 scale-100"
-                leave="transition ease-in duration-75"
-                leaveFrom="transform opacity-100 scale-100"
-                leaveTo="transform opacity-0 scale-95"
-            >
-                <Menu.Items
-                    className="absolute right-2 z-10 mt-1 tablet:min-w-[60px] origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
-                    <div className="py-1 ">
-                        {
-                            milestoneMenus.map(v =>
-                                <Menu.Item key={v.name}>
-                                    {({active}) => (
-                                        <a
-                                            href="javascript;"
-                                            onClick={(e) => {
-                                                e.preventDefault();
-                                                v.onClickHandler();
-                                            }}
-                                            className={classNames(
-                                                active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
-                                                'block px-4 py-2 tablet:text-[16px] mobile:text-sm'
-                                            )}
-                                        >
-                                            {v.name}
-                                        </a>
-                                    )}
-                                </Menu.Item>
-                            )
-                        }
+        authInfo?.milestoneAuth ?
+            // <div className="flex-shrink-0 pr-2 overflow-visible">
+            (
+                <Menu as="div" className="self-start flex-shrink-0 pr-2 text-center">
+                    {/*<Menu as="div" className="relative inline-block text-center">*/}
+                    <div>
+                        <Menu.Button
+                            className="inline-flex h-8 w-8 items-center justify-center rounded-full bg-transparent bg-white text-gray-400 hover:text-gray-500 focus:outline-none">
+                            <span className="sr-only">마일스톤 메뉴</span>
+                            <IoEllipsisVertical className="h-5 w-5" aria-hidden="true" data-role='milestone-menu'/>
+                        </Menu.Button>
                     </div>
-                </Menu.Items>
-            </Transition>
-        </Menu>
-    );
+                    <Transition
+                        as={Fragment}
+                        enter="transition ease-out duration-100"
+                        enterFrom="transform opacity-0 scale-95"
+                        enterTo="transform opacity-100 scale-100"
+                        leave="transition ease-in duration-75"
+                        leaveFrom="transform opacity-100 scale-100"
+                        leaveTo="transform opacity-0 scale-95"
+                    >
+                        <Menu.Items
+                            className="absolute right-2 z-10 mt-1 tablet:min-w-[60px] origin-top-right rounded-md bg-white shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none">
+                            <div className="py-1 ">
+                                {
+                                    milestoneMenus.map(v =>
+                                        <Menu.Item key={v.name}>
+                                            {({active}) => (
+                                                <a
+                                                    href="javascript;"
+                                                    onClick={(e) => {
+                                                        e.preventDefault();
+                                                        v.onClickHandler();
+                                                    }}
+                                                    className={classNames(
+                                                        active ? 'bg-gray-100 text-gray-900' : 'text-gray-700',
+                                                        'block px-4 py-2 tablet:text-[16px] mobile:text-sm'
+                                                    )}
+                                                >
+                                                    {v.name}
+                                                </a>
+                                            )}
+                                        </Menu.Item>
+                                    )
+                                }
+                            </div>
+                        </Menu.Items>
+                    </Transition>
+                </Menu>
+            )
+            : null
+    )
+        ;
 }
 
 export default MilestoneCardMenu;
